Remove dependency from config repo
The templates infrastructure replaces this
