+++ charger/template_test.go
@@ -8,6 +8,26 @@ import (
 	"github.com/thoas/go-funk"
 )
 
+var acceptable = []string{
+	"invalid plugin type: ...",
+	"missing mqtt broker configuration",
+	"mqtt not configured",
+	"invalid charger type: nrgkick-bluetooth",
+	"NRGKick bluetooth is only supported on linux",
+	"invalid pin:",
+	"hciconfig provided no response",
+	"connect: no route to host",
+	"connect: connection refused",
+	"error connecting: Network Error",
+	"i/o timeout",
+	"recv timeout",
+	"(Client.Timeout exceeded while awaiting headers)",
+	"can only have either uri or device", // modbus
+	"sponsorship required, see https://github.com/evcc-io/evcc#sponsorship",
+	"eebus not configured",
+	"unexpected status: 400", // easee
+}
+
 func TestChargerTemplates(t *testing.T) {
 	test.SkipCI(t)
